Refactor path segments handling
