+++ components/HighlightProvidedSelection.tsx
@@ -4,78 +4,8 @@ import useDeepCompareEffect from 'use-deep-compare-effect';
 import { isUnderContentContainer } from './Container';
 import { fragmentToRangeList, selectionToFragment } from './lib/deeplinks';
 import { selectRanges, SelectRangesOptions } from './lib/selectranges';
+import { getPathSegments } from './pathsegments';
 import { useRenderState } from './renderstate';
-
-type SelectionPathParts = {
-  basePath?: string;
-  pathPrefix?: string;
-  selectionId?: string;
-};
-
-/**
- * Parses a pathname into parts used to highlight selections.
- * basePath: The base path for the siet. Only supports a single level base path.
- * pathPrefix: The prefix that identifies that type of selection. Must start with '_'
- * selectionId: The content selection id
- */
-function getSelectionPathParts(pathname: string): SelectionPathParts {
-  // Only handle deeplink paths (identified by the presence of ':')
-  // to allow default behavior for all other fragments such as anchor links
-  let basePath, pathPrefix, selectionId;
-  const pathParts = pathname.split('/').filter((part) => part);
-  const partCount = pathParts.length;
-  let part = pathParts.pop();
-  switch (partCount) {
-    case 0:
-      // No parts present
-      break;
-    case 1:
-      // /<selection id>
-      // /<base path>
-      if (part?.includes(':')) {
-        selectionId = part;
-      } else {
-        basePath = part;
-      }
-      break;
-    case 2:
-      // /<base path>/<selection id>
-      // /<path prefix>/<selection id>
-      if (part?.includes(':')) {
-        selectionId = part;
-        const baseOrPrefix = pathParts.pop();
-        // path prefixes must start with an underscore
-        if (baseOrPrefix?.startsWith('_')) {
-          pathPrefix = baseOrPrefix;
-        } else {
-          basePath = baseOrPrefix;
-        }
-      } else {
-        // <unknown format>
-        // Preserve base path, ignore remainder of path
-        basePath = pathParts.pop();
-      }
-      break;
-    case 3:
-      // /<base path>/<path prefix>/<selection id>
-      if (part?.includes(':')) {
-        selectionId = part;
-        pathPrefix = pathParts.pop();
-        basePath = pathParts.pop();
-      }
-      // <unknown format>
-      break;
-    default:
-      console.log('Unsupported path format');
-      break;
-  }
-
-  return {
-    basePath,
-    pathPrefix,
-    selectionId,
-  };
-}
 
 export function HighlightProvidedSelection(props: {
   options?: SelectRangesOptions | undefined;
@@ -87,12 +17,12 @@ export function HighlightProvidedSelection(props: {
   const ranges = renderState.ranges;
 
   useDeepCompareEffect(() => {
-    const pathParts = getSelectionPathParts(window.location.pathname);
+    const pathSegments = getPathSegments(window.location.pathname);
 
-    if (pathParts?.selectionId) {
+    if (pathSegments?.selectionId) {
       const selectionRanges = fragmentToRangeList(
         document,
-        pathParts.selectionId
+        pathSegments.selectionId
       );
       selectRanges(selectionRanges, props.options);
       return;
@@ -115,9 +45,7 @@ export function HighlightProvidedSelection(props: {
       return;
     }
 
-    const { basePath, pathPrefix, selectionId } = getSelectionPathParts(
-      window.location.pathname
-    );
+    const { basePath, pathPrefix } = getPathSegments(window.location.pathname);
 
     // Based on the fragment in the original deeplinks.ts library.
     const timeoutHandle = setTimeout(() => {
